MER-151 finalne poprawki do review

+++ vulcanus/src/main/java/com/merkury/vulcanus/model/entities/Comment.java
@@ -24,7 +24,7 @@ public class Comment {
 
     private String text;
     private Double rating;
-    private long likes;
+    private Integer likes;
 
     @ManyToOne
     @JoinColumn(name = "spot_id")
